% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{mipalette}
\alias{mipalette}
\title{MI Umbrella commonly used color palettes}
\usage{
mipalette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{MIStartMap}, \code{OldMIStartMap},  \code{MISafeStart},
\code{CDCTransmissionLevels}, \code{MISMTransmissionLevels},  \code{CDCCommunityLevels}, \code{MISMCommunityLevels},
\code{MISMTequilaSunrise},  \code{USRegionalTrends} , \code{RetroRainbow} ,
\code{RetroRainbow2}, \code{SunForest6}, \code{BeachBall}, \code{GreyTealGold},
\code{BigTop}, \code{BigTop2}, \code{Midsummer}, \code{SummerHike}}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are various color palettes that are used frequently for the MI Umbrella projects. They include color schemes from MI Start Map, MI Lighthouse, and various COVID levels.
This function was borrowed pretty much directly from Kartik Ram's wesanderson package: \url{https://github.com/karthik/wesanderson}
}
\examples{
mipalette("MIStartMap")
mipalette("MISMCommunityLevels")
mipalette("SunForest6")
mipalette("SunForest6", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- mipalette(21, name = "RetroRainbow", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
