% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{mipalettes}
\alias{mipalettes}
\title{Complete list of commonly used color palettes by MI Umbrella}
\format{
An object of class \code{list} of length 23.
}
\usage{
mipalettes
}
\description{
Use \code{\link{mipalette}} to construct palettes of desired length.
}
\keyword{datasets}
